CX360/requirements.txt
fastapi
uvicorn
mysql-connector
boto3
pydantic
findspark
pyspark
pandas
python-multipart
pymysql